fix: changed class structure and access modifiers for methods and variables
